framework: allow storage buffer on vkdf-descriptor methods
As far as I understand, that (at least most) code would be the same
for UNIFORM_BUFFER and STORAGE_BUFFER, so we could reuse it.

Additionally, it is more consistent with other methods on the same
header, like vkdf_create_descriptor_pool, that already receives the
type

Now instead of passing is_dynamic, we pass directly the descriptor
type.