ssbo: generate 4 shaders based on a template
The shaders are slightly different depending on the number of
components we want to use. The alternative to this patch would be
implement the template logic on the test itself.

But although this option leads to having 4 extra shader files around,
it allows to keep glslangValidator as a external program, and maintain
the logic on the test simpler.