ssbo16: ensure buffers are multiple of 4
From vk spec, vkCmdFillBuffer documentation:

    size is the number of bytes to fill, and must be either a multiple
    of 4, or VK_WHOLE_SIZE to fill the range from offset to the end of
    the buffer. If VK_WHOLE_SIZE is used and the remaining size of the
    buffer is not a multiple of 4, then the nearest smaller multiple
    is used.

So we set the size of the ssbo/vgo to be a multiple of 4, adding a
padding if necessary, in order to ensure that we get the last
values. This affects 1 and 3 component execution

v2: include vbo too
v3: fix formula. vbo doesn't really need it, but I keep it, as
    makes sense