framework: mark dynamic objects
We can use this info to optimize rendering of static objects because
we know their transforms stay fixed and we won't need to update
them. This allows us to map scene resources per tile for all
static objects only once so we never have to update any of them again
during rendering.