framework/object: allow objects to specify if they can cast or receives shadows
We will use this in the scene renderer: objects that cannot cast shadows will
not be renderered to the shadow map. Also, the object UBO provided to scene
clients will provide info on whether each instance can receive shadows so when
apps call the compute_lighting() function they get the expected behavior
automatically.