framework/scene: honor object lighting behaviors
Only render to the shadow map objects that can cast shadows and include
whether the objects can receive shadows in the static object UBO.

This change means that we need a new UBO for the shadow map, since
not all objects in a visible tile will necessarily be rendered into
the shadow map, depending on their shadow casting behavior.

Also, fix a bug where instance size of the object UBO was not correct.