data/lighting: optimize shadow map sampling
Check if the current fragment is outside the light's area of influence
(outside the shadow map) and in that case return early marking the
fragment as being fully in the shadow without sampling the shadow map
(which would be incorrect anyway).

This not only makes sure that we shadowe fragments outside the light's
are of influence, it also prevents expensive shadow map samples for such
fragments, which is particuarly relevant when PFC is enabled.