framework/scene: improve spotlight (sub)tile culling
So far we were only culling at tile level, now we also consider
subtiles like we do for scene rendering. In order to do this, we
need shadow caster counts per subtile too, so instead of
computing them on the fly when we create the UBO for the shadow
map we extend the object count algorithm to also consider shadow
caster counts.

Also, reuse more of the existing infrastrucutre for testing
(sub)tile visibility for the case of spotlights.