scene/framework: further improve spotlight culling
We can test visible tiles against the actual cone of the spotlight
for more strict visibility. However, precission differences between
CPU and GPU make this a bit unreliable and some times we can end
up with a small part of an object in the light of the cone and still
consider that its tile is not visible. We can fix that by adding some
error margin to the visibility test, but that also reduces the
effectivess of the test, so do not enable this for now.