framework/light: move the dirty flag from the scene light to the light
Apps will interact with the light object, so that's where we want to
track dirty status for a light. If the light is used in a scene, then
the scene can also set the dirty flag of the light object and it is
responsible for setting it to false when it is done processing light
updates. If the light is not added to a scene object, then it is
the app's responsibility to reset the dirty flag of a light.