framework/light: drop 'is_dynamic' field
This was intended to be used by the scene object to skip updates
on static lights, but it is not worth it at the moment since the
scene will go through all the lights in the scene to check their
dirty flags so skipping that check with another check that is
as expensive doesn't make sense.

In the future we might want to sort the list of lights so that
all static lights go first. Then the scene object could use the
number of static lights to skip all of them in one go when checking
for dirty lights.