framework/camera: improve dirty state management
Now we can cache things such as the frustum or the view matrix so we
can reuse it if they have not been invalidated.

Removed the 'frustum at distance' functions for simplicity, since we
are not currently using them and won't be cached. We can add them back
later when we need them.