framework/scene: have the scene object define the render passes
Until now the application was setting the one defining the single
render pass in use, however, we want the scene to have control over
this so it can define and use different render passes depending
on what needs to render.

This makes the renderpass begin callback unnecessary and allows us
to move the render target definition to the scene object too,
simplifying client implementations.

Besides this, the scene can now use different render passes for
static and dynamic geometry so we don't have to rebuild the static
primary command buffer on every frame due to dynamic updates. To,
achieve this, the record commands callback takes the command into
which it needs to record, so it doesn't have to decide if it is a
primary or a secondary, that is decided by the scene object.

The scene object will now generate frames in color attachment
optimal layout, and vkdf_command_buffer_create_for_present()
will assume this layout and use a barrier to transition the color
image to the transfer source layout.

Clients should specify the clear colors to use before preparing
the scene so appropriate render passes can be created. Color
clear is optional but we always require a depth clear.