framework/scene: move presentation to the scene object
The scene now takes care of copying the resulting image to the
swapchain for repsentation, which simplifies clients.

Clients can still do postprocessing of the scene rendering
by providing a callback that will be executed after scene rendering
and before presentation. The callback receives a wait semaphore
(for scene rendering) and a draw semaphore (for presentation to
wait on), so that postprocessing can be properly synchronized.

Note that We are not handling fencing correctly here: we should
introduce the fence in the submission of the presentation command,
but we were not doing that before either...