framework/scene: synchronize presentation and rendering properly
We want a fence to know when we are done copying the previous frame
to the swapchain, as that signals the availability of the render
target for rendering the new frame.

Because we always need to wait for rendering to complete before
presenting, with the above fence in place we can now submit
resource updates for the current frame without waiting and
postpone the wait on the fence to the point in which we are
about to start rendering the new frame.