framework/scene: rename threading data used for tile processing
So far have only been using threads for processing tiles for static
geometry. The thread data used for this needs to be persistent
because it is used between frames to tell if the list of visible
tiles has changed and we need to rebuild the primary command buffer.

Since we want to use thread tasks for other purposes, we will need
other thread data instances, so naming them specifically for their
intended use is relevant.