framework/scene: thread shadow map updates
We create a thread job per shadow caster light that takes care of
testing if the shadow map is dirty and recording its update if needed.

At the moment all threads that require to record a new shadow map
will use the same command buffer pool, which is not safe since we
are not locking. To fix this though, we need to update the thread
job interface so we know the ID of the thread that is running the
job. Will fix that in the next commit.