framework/scene: command buffer freeing fixes
Sometimes we were attempting to free a command buffer using a different
pool than the one that was used to crate it, which is not correct.
Specifically, for tile jobs, we want to use the job-id instead of the
thread-id to select pools so that a particular job range always uses
the same pool, no matter the thread that is running it. This is safe
because we split the jobs so that there is exactly one job per thread,
so there is no chance that two different threads end up using the
same pool in the same iteration.

Also, we were not freeing obsolete shadow map command buffers. Didn't
notice this before because the validation layers won't report this since
the pools are destroyed by the end of the execution and that frees
all command buffers allocated from them.

Finally, made new_inactive_cmd_buf() more generic so it makes sense to
use it to free any command buffers generated by the scene. Previously
the interface required to pass a TileThreadData pointer which is a
structure associated with tiling command buffers only.