framework/util: add means to compute a view matrix for a given view direction
We will use this to compute the view matrix for directional lights.

This implementation assumes a viewer located at the origin, which
is handy for the implementation of directional lights where we want
the viewer positioned at the  camera's location, not the light. We
can achieve this by taking the result of this function and applying
a translation to position it where we need.