Revert "framework/scene: synchronize resource updates with frame rendering"
This reverts commit de7d83011c91d74d552f90631850dda7a7b8d0b1.

That was not done properly since we were trying to wait on a semaphore
that was consumed by the presentaiton job of the previous frame. To do
this right we would need to have another semaphore or have the last
rendering job signal a fence and then wait on that.