demos/spoza: add a clipping option
The sponza object is pretty big, but being a single object the
scene renderer will either clip it completely or not at all, which
is not great.

Handle it as a dynamic scene object instead. This will give us a
chance to mesh clipping on every frame instead.

Also, disable validation layers by default, since they have quite
an effect on performance with this demo.