framework/scene: respect minUniformBufferOffsetAlignment
Our shadow map data comes right after the light data in the UBO, so
make sure that we respect the offset alignment requirement for
the starting offset of the shadow map info.

Fixes shadow mapping issues in the Sponza demo after recent changes
in the Intel driver to support push UBOs (which bumped alignment
requirements from 16B to 32B).