framework/pipeline: make vkdf_create_gfx_pipeline more flexible
Let it specify more than one color attachment (to initialize blend
state) and accept a NULL fragment shader for pipelines that don't
need one.

This simplifies pipeline creation in the sponza demo for
the depth pre-pass and deferred rendering modes. Also, it removes
the validation error in the depth-prepass with opacity pipeline
in this demo because now we can state that the pipeline doesn't use
any color attachments.