framework/camera: track if the camera has moved
We track if the camera's position has changed since the last time
the user called vkdf_camera_set_dirty(cam, false).

We expect this to be called by the end of the previous frame, in
which case we are testing if the camera's position has changed
since the last frame.