framework/scene: reset static primary commands instead of freeing
This has some visible performance gains when we have a lot if tiles
in the scene and we have only a couple of these anyway.

We need to be careful not to reset a command buffer when it is still
being used for rendering, so we use a list of command buffers and we
cycle between them to ensure this never happens. As of now, we always
wait on presentation of frame N-1 before we render frame N, so a list
of 2 command buffers is sufficient.