framework/scene: provide an option to never free secondary command buffers
This is the most performant option. If memory is a concern, maybe if there
is a very large number of tiles, we can always fall back to disabling this
and play with the cache size instead.

Not freeing is now the default option. To disable set SCENE_FREE_SECONDARIES
to true (needs recompile).