framework/scene: rework shadow map updates to improve batching
Previously, we recorded multiple command buffers so we can use
multiple threads to record the updates, however, it is unlikely
that we have enough dirty shadow maps in a frame that this will
pay off. Batching all shadow map updates in a single command buffer
looks like a better option.