framework/image: improve image loading
Sometimes SDL doesn't provide accurate format data, so don't rely on that.

Also, make our loader a bit more flexible so it can handle more than
RGB8_UNORM images. The sponza demo features a specular texture for the
walls that is only 8bpp with a single channel representing intensity
while other specular maps are RGB grayscale data. Now we load this
properly and fix specular reflections on the walls of the Sponza model.

Finally, we make a separate internal helper to create the image from
the pixel data so we can reuse it to create images from data that doesn't
come only from image files. We will use this in the implementation
of new features.