demos/sponza: fix validaton error on forward rendering pipeline
Due to the pipeline layouts we use, we were bounding sets 0..2 and 4
just once, then for each mesh we were binding a different descriptor
at set 3 for every mesh with the textures. The validation layers
complained that after this, set 4 was not bound.

Change the forward and deferred layouts so that set 4 (the shadow
map sampler) becomes set 3 and we always bind sets 0..3 once. The
shadow map sampler set is not used in the gbuffer pass, but the
shaders just ignore it. This allows us to share the same pipeline
layout for both deferred and forward and get rid of the validation
complaints. Ideally, we should just have different layouts for
forward and deferred.