demos/sponza: fix sky color in deferred rendering
The sky color is really the clear color, only that in deferred
we don't clear and instead we ouput the clear color in the shader
directly when we detect that the pixel was not written to in the
GBuffer pass. To check for this, we used to rely in the eye-space
light position vector that we stored in the GBuffer, but since we
have now removed this with the previous patch, we need to implement
a different check. Now we check against the GBuffer normal.