framework/scene: enable/disable light shadows based on the shadow spec
Previously we expected that applications would create a light, then
enable shadows for it and then add it to the scene with a shadow
specification. The scene would assert if we received a shadow spec
for a light with shadows disabled or viceversa.

Just make things easier and automatically enable/disable shadows for a
light depending on whether we get a shadow specification for it when
it is added to the scene.

Adapt demos to honor this.