demos/sponza: make the shadow map box tighter
For slightly improved shadow mapping quality.

If we make it smaller than this then the shadow map won't cover
the entire area and that means that when we are in one end of the
Sponza model, looking at the other and moving backwards we can
see the shadow mapping disappearing on the other end of the model.

If we want to increase quality further without leading to this
artifacts we need to implement support for cascaded shadow maps.