framework/scene: compute shadow box properly for directional lights
So far we have been taking a shortcut to implementing shadow mapping
for directional lights by centering the shadow box at the camera and
computing the size of the box in world space. What we should be doing
instead is:

1. Compute the dimensions of the shadow box in light space
2. Center the shadow box at the center of the bounding box that encloses
   the camera's frustum.

Because of 2), now the shadow box changes not only when the camera moves
but also when it rotates, so we need to update the shadow map in this
case too.

We also provide the application with new shadow spec parameters to
scale the size of the shadow box in any axis and to offset its center
along the viewing direction of the camera to have greater control of
the area covered by the shadow map. These new parameters can be
used to configure tighter shadow boxes depending on the characteristics
of the scene, which can improve shadow mapping quality.