framework/init: rework device extensions
Now we can specify if the extensions are optional or required and
we fail inmediately if a required extension is not available. we
can also query which extensions have been eventually enabled after
creating the logical device.

Also added VK_KHR_maintenance1 to the list of required extensions.
We are not using anything specific that needs this at present, but
since this extension incorporates API fixes it makes sense to
require it going forward.