framework/scene: always use linear RGBA format for the color render target
And blit to the presentation surface to convert to the presentation format
if necessary.

The theory here is that we want all our rendering to happen in linear
RGBA first, and then, if the presentation surface is sRGB, we
convert our linear RGBA color buffer to sRGB as we copy to the
presentation image via a blit operation.