framework/scene: fix SSAO halo artifacts wit blur pass
These artifacts are produced when the blur pass averages SSAO
values for pixels that belong to different pieces of geometry
that happen to have very different occlusion factors. In that
case, the blur proccess will "leak" occlusion between the edges
of the geometries.

In the Sponza demo, this can be observed clearly around the edges
of the curtains near the top of the pillars. The edges of the
curtains have low occlusion values, whereas the pillars are highly
occluded. The result is that the pixels in the edge between the
curtain and the pillar get 50% occlusion value, making them stand
out as a halo.

To reduce this, we avoid blurring together pixels that have very
different occlusion factors. The threshold is configurable. If
the threshold is too low, the blur won't eliminate the SSAO noise,
and it it is too high, it will produce the halo effect.

This doesn't completely elminate the halos, but it is cheap and
helps enough to remove most of the problem. A better fix would
probably be to sample the depths of the pixels and only blur
them together if their depths are similar enough (without looking
at the difference between their SSAO values), but that would be
more expensive due to the extra depth lookups and for now this
seems to work well enough.

We also switch to using nearest filtering for the SSAO texture in
the Sponza demo to further avoid occlusion leaking due to
filtering.