framework/model: honor sRGB flag also for specular maps.
Since this are also color textures.

However, if when we load the image we find it isn't RGB or
RGBA, we ignore the sRGB flag. This is because in that
case the texture is less likely to represent color data
but something else such as specular intensity, which is not
sRGB encoded.

Also, at least on Intel/Mesa, we are abusing our blits to
generate mipmaps since we are ignoring the fact that the driver
doesn't allow to blit to RGB destinations. We have been getting
away with that so far without issues, but trying to blit to a
sRGB destination with less than 3 components results in a GPU
hang.