framework/scene: allow application to decide the size of the framebuffer
So far we have been using a framebuffer that exactly matched the dimensions
of the window.

Since now the framebuffer size can be different from the presentation size,
allow applications to specify the filter to be used during presentation.

Also, use this feature to trivially implement supersample antialiasing
(SSAA) in the Sponza demo (disabled by default)