framework/scene: present from the image provided by the post-processing
Even if we had a post-processing callback set we were still always presenting
from the scene renderer color buffer, which in practice meant that our
post-processing had to render to it.

With this change, post-processing callbacks can return the VkImage
that contains the final rendering so the scene renerer can present
from it.