framework/scene: discard pixels in the SSAO blur pass based on depth distance
Do the right thing and use the depth difference between neighbours
to discard blurs, instead of relying on distance between the SSAO values.
This adds an extra cost due to the additional lookups, but it produces much
better results.

Notice that this won't eliminate artifacts caused by downsampling
the SSAO pass, since in that case sampling from the SSAO texture
requires filtering, and we get averages (linear) or just the closest
value in the downsampled texture (nearest), so some artifacts are
expected in any case.