framework/scene: use a specialization constant for the SSAO sample kernel size
This allows the shader compiler to produce better (optimal) code which can
lead to a significant performance boost.

This required that we added a new overload of the gfx pipeline constructor
to accept VkPipelineShaderStageCreateInfo instead of VkShaderModule for the
shaders to use, so the caller can provide constant specialization info when
needed.