framework/scene: rework post-processing chain
Our previous solution was not valid. We need to know the inputs
and outputs of each post-processing stage ahead of rendering time
so we can create proper descriptor sets for each post-processing
stage.

The above means we cannot rely on the application's postprocessing
callback to return the output image since that is called at render time.
Instead, we remove the postprocessing callback hook from
vkdf_scene_set_scene_callbacks() and we require that is added via the
new API call vkdf_scene_enable_postprocessing(), which also takes a pointer
to the last image produced by the post-processing chain. With that
information we can properly setup inputs and outputs in the scene's
post-processing chain for all stages (FXAA only for now).