framework/scene: provide means to reduce the rate of shadow map updates
Until now we would update dirty shadow maps on every frame. Now we
provide an option to only update dirty shadow maps every N frames
(or even never after the first time).

This is useful in small scenes with directional lights where the
initial shadow box encloses the entire scene. In this case, updating
the shadow map as the camera moves isn't really needed and besides
wasting GPU resources it can add shadow flickering artifacts.
This commit also updates the Sponza demo to never update the
shadow map, since it falls in this scenario.

For scenes with dynamic objects that are in the view field of
shadow casting lights and are dirty (that is, they have been
animated since the previous frame), we still need to record the
shadow map and we can't skip frames, since otherwise we can get
self-shadowing artifacts on the objects.