framework/model: add option to limit collision testing to specific meshes
This is useful when some meshes in a model are setup in a way that box
collision won't do (such as very irregular meshes with holes, etc), or
when there are a lot of small meshes that don't really matter for
collision testing.

While we don't implement collision detection here, we will be using
this soon for scene-level collision detection.