framework/light: add a helper to obtain the light's position only
Previously, the helper would return a vec4 where .w would include the
light's type. This renames the helper to make this clear and adds a
new one that only returns a vec3 with the position.

All uses of the old helper where consuming a vec3 only (with the
type conversion happening implicitly) so we don't need to change
any users of the old version.