framework/scene: make the scene compute and update eye-space light data
We compute compute light position and/or direction (depending on light
type) in eye-space as needed and put it in the scene light data UBO
for applications to use.

A boolean flag controls whether the scene will compute eye-space data
for lights. This will be set to true automatically for deferred
rendering.

Update the sponza demo to take advantage of this.