framework/scene: add scene objects representing light volumes
This will be useful to optimize the lighting pass in deferred
renderers.

We create and add the objects to the scene automatically when the
lights are added and we update the volume object properties when
the lights are dirty. By adding the volumes to the scene we get
light clipping for free.

Because these do not represent real objects we disable collision
testing against them.

We store these volume objects with specific object set identifiers
so applications can recognize them in scene callbacks and handle
them properly. For example, a typical deferred renderer would not
render these objects to the gbuffer and would instead render them
during the lighting pass to optimize lighting calculations.