framework/light: add a helper to get the scale of a light's volume
In order to properly define a 3D geometry representing a light's
volume, like with any other 3D object, we need to know its position,
its orientation and its scale. We know the first two since they
are parameters of the light source, but the scale has to be computed
from the light's intensity and attenuation factors.