framework/scene: add a private data field to the object UBO
When we render light volume objects for the lighting pass we will
need a way to identify the index in the light UBO of the volume
object we are rendering. The scene will use this private object
field to do that. This is a bit of a hack, the right way to
implement this feature would be to have VkdfSceneObject inheriting
from VkdfObject and add this information there (like we do for
lights with VkdfSceneLight), but for now we take the easy way out.

This requires that we update many shaders that access object UBO
data provided by the scene. In the future we probably want to
include the scene object data structure from a GLSL file to avoid
this.