framework/scene: add API to remove a light source
We have to be a bit careful here, since removing a light source
changes the indices of the lights in the vector, which forces
us to re-upload the light and shadow map UBOs, since their
offsets in the UBO are computed based on the light index. This
also means that we need to re-index volume objects, since they
store the index of the light the light they are associated to
in the objects's private data.

We were only using the boolean has_shadow_caster_lights to know
at prepare time if we had shadow casting lights in the scene
and skip shadow mapping setups if we didn't, but since we are
moving to a more dynamic implementation where we may add or
remove lights from the scene at any moment, this doesn't make sense
any more.