framework/object: flag whether to do mesh-level collision testing
If set to true, collision testing will happen against the bounding
boxes of individual meshes in the object, otherwise it will take
place against the single object's bounding box.

Enable this in the sponza demo, since there we have a single object
and we really want to check for collision against individual meshes
for things to work.