framework/scene: support dynamic materials for dynamic objects
Until now we only uploaded all materials once, in one go, and assumed they
would remain the same. Now, we keep track of whether materials for any model
are dirty and upload them dynamically as needed when we process dirty object
state.

Notice that static objects only get their copies of the materials
uploaded once to the static material UBO, so material updates won't
affect static objects, even if they share the underlying model instance.

At present, we only mark materials as dirty if new materials are added,
not if they are modified, since we don't have API for that. We will add
that later on.