framework/init: do not enable optional device extensions
This change avoids a validation layer warning when using
VKDF_NO_SWAPCHAIN, as in this case we are not initializing the
instance extension VK_KHR_SURFACE. By spec, VK_KHR_swapchain depends
on VK_KHR_surface, so if we enable the swapchain we need to enable
VK_KHR_surface.

Alternatively, we could enable VK_KHR_surface is available.

v2: fix indentation