framework/buffer: align buffer size for map/unmap to honor memory requirements
To prevent Validation Layer error VUID-VkMappedMemoryRange-size-01390.

Right now vkdf-create-buffer is allocating what the buffer really
requires. So for example, if you are trying to allocate 96 bytes, but
the specific driver allocates based on an aligment of 256, the
allocate size will be 256.

So then, if you try to fill up the whole buffer using an array of 96,
if you map/unmap only the size, you would get the validation error
below. This commit maps/unmaps using an aligned size. The copied
memory still uses the original size.

Full text of the validation error:

VKDF ERROR: validation layer: [ VUID-VkMappedMemoryRange-size-01390 ]
vkFlushMappedMemoryRanges: Size in pMemRanges[0] is 0x60, which is not
a multiple of VkPhysicalDeviceLimits::nonCoherentAtomSize (0x100). The
Vulkan spec states: If size is not equal to VK_WHOLE_SIZE, size must
either be a multiple of VkPhysicalDeviceLimits::nonCoherentAtomSize,
or offset plus size must equal the size of
memory. (https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/vkspec.html#VUID-VkMappedMemoryRange-size-01390)