framework: add full wayland support on sdl platform
In this case it tries to use the surface using
vkCreateWaylandSurfaceKHR instead of vkCreateXlibSurfaceKHR.

Note that now we are defining both VK_USE_PLATFORM_XLIB_KHR and
VK_USE_PLATFORM_WAYLAND_KHR, so the compiled vkdf would support
both. This is useful (and needed) for platforms that supports both,
and we want to force one (like using SDL_VIDEODRIVER=wayland)

But for environments that we only use one, we could set just
one. Perhaps passing a new parameter to autogen.sh